+++ topology.go
@@ -23,6 +23,7 @@ type Topology struct {
 	rawOutput bool
 	upch      chan string
 
+	metrics   MetricsClient
 	invalid   [LogLineNumFields]int64 // count validation errors (by field)
 	malformed int64                   // count parse or empty records
 
@@ -57,6 +58,19 @@ func NewTopologyFromConfig(cfg *Config) (*Topology, error) {
 		},
 	}
 
+	// Create the metrics client first since it's injected into components parameters.
+	if cfg.Metrics.Name != "" {
+		tp.metrics, err = cfg.Metrics.desc.New(cfg.Metrics.DecodedConfig)
+		if err != nil {
+			return nil, fmt.Errorf("error creating metrics interface: %q: %v", cfg.Metrics.Name, err)
+		}
+	}
+
+	// Assign a dummy client if no one was installed
+	if tp.metrics == nil {
+		tp.metrics = NopMetrics{}
+	}
+
 	// * Create input
 	inCfg := InputParams{
 		ComponentParams{
@@ -65,6 +79,7 @@ func NewTopologyFromConfig(cfg *Config) (*Topology, error) {
 			FieldName:      cfg.fieldName,
 			CreateRecord:   cfg.createRecord,
 			ValidateRecord: cfg.validate,
+			Metrics:        tp.metrics,
 		},
 	}
 	tp.Input, err = cfg.Input.desc.New(inCfg)
@@ -81,6 +96,7 @@ func NewTopologyFromConfig(cfg *Config) (*Topology, error) {
 				FieldName:      cfg.fieldName,
 				CreateRecord:   cfg.createRecord,
 				ValidateRecord: cfg.validate,
+				Metrics:        tp.metrics,
 			},
 		}
 		fil, err := cfg.Filter[idx].desc.New(filCfg)
@@ -111,6 +127,7 @@ func NewTopologyFromConfig(cfg *Config) (*Topology, error) {
 				FieldName:      cfg.fieldName,
 				CreateRecord:   cfg.createRecord,
 				ValidateRecord: cfg.validate,
+				Metrics:        tp.metrics,
 			},
 			Index:  i,
 			Fields: tp.outFields,
@@ -162,6 +179,7 @@ func NewTopologyFromConfig(cfg *Config) (*Topology, error) {
 				FieldName:      cfg.fieldName,
 				CreateRecord:   cfg.createRecord,
 				ValidateRecord: cfg.validate,
+				Metrics:        tp.metrics,
 			},
 		}
 		tp.Upload, err = cfg.Upload.desc.New(upCfg)